# openapi/personnel-administration-api.yaml
openapi: 3.0.3
info:
  title: B4RRHH Personnel Administration API
  version: 0.1.0
paths:
...
  /rule-systems/{ruleSystemCode}/employees/{employeeNumber}:
    get:
      operationId: getEmployeeByBusinessKey
      summary: Get employee by business key
      parameters:
        - in: path
          name: ruleSystemCode
          required: true
          schema:
            type: string
            maxLength: 5
        - in: path
          name: employeeNumber
          required: true
          schema:
            type: string
            maxLength: 15
      responses:
        '200':
          description: Employee found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeResponse'
        '404':
          description: Rule system or employee not found
  /rule-systems/{ruleSystemCode}/employees/{employeeNumber}/presences:
    post:
      operationId: createPresenceByBusinessKey
      summary: Create employee presence by business key
      parameters:
        - in: path
          name: ruleSystemCode
          required: true
          schema:
            type: string
            maxLength: 5
        - in: path
          name: employeeNumber
          required: true
          schema:
            type: string
            maxLength: 15
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePresenceRequest'
      responses:
        '201':
          description: Presence created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PresenceResponse'
        '400':
          description: Invalid request
        '404':
          description: Rule system, employee, or catalog not found
        '409':
          description: Presence business conflict
    get:
      operationId: listEmployeePresencesByBusinessKey
      summary: List employee presences by business key
      parameters:
        - in: path
          name: ruleSystemCode
          required: true
          schema:
            type: string
            maxLength: 5
        - in: path
          name: employeeNumber
          required: true
          schema:
            type: string
            maxLength: 15
      responses:
        '200':
          description: Presence list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PresenceResponse'
        '404':
          description: Rule system or employee not found
  /rule-systems/{ruleSystemCode}/employees/{employeeNumber}/presences/{presenceId}:
    get:
      operationId: getPresenceByIdByBusinessKey
      summary: Get employee presence by id and business key
      parameters:
        - in: path
          name: ruleSystemCode
          required: true
          schema:
            type: string
            maxLength: 5
        - in: path
          name: employeeNumber
          required: true
          schema:
            type: string
            maxLength: 15
        - in: path
          name: presenceId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Presence found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PresenceResponse'
        '404':
          description: Rule system, employee, or presence not found
  /rule-systems/{ruleSystemCode}/employees/{employeeNumber}/presences/{presenceId}/close:
    post:
      operationId: closePresenceByBusinessKey
      summary: Close employee presence by business key
      parameters:
        - in: path
          name: ruleSystemCode
          required: true
          schema:
            type: string
            maxLength: 5
        - in: path
          name: employeeNumber
          required: true
          schema:
            type: string
            maxLength: 15
        - in: path
          name: presenceId
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClosePresenceRequest'
      responses:
        '200':
          description: Presence closed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PresenceResponse'
        '400':
          description: Invalid request
        '404':
          description: Rule system, employee, or presence not found
        '409':
          description: Presence already closed
  /employees/{employeeId}/presences:
    post:
      operationId: createPresence
      summary: Create employee presence
      parameters:
        - in: path
          name: employeeId
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePresenceRequest'
      responses:
        '201':
          description: Presence created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PresenceResponse'
        '400':
          description: Invalid request
        '404':
          description: Employee or catalog not found
        '409':
          description: Presence business conflict
    get:
      operationId: listEmployeePresences
      summary: List employee presences
      parameters:
        - in: path
          name: employeeId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Presence list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PresenceResponse'
        '404':
          description: Employee not found
  /employees/{employeeId}/presences/{presenceId}:
    get:
      operationId: getPresenceById
      summary: Get employee presence by id
      parameters:
        - in: path
          name: employeeId
          required: true
          schema:
            type: integer
            format: int64
        - in: path
          name: presenceId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Presence found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PresenceResponse'
        '404':
          description: Presence not found
  /employees/{employeeId}/presences/{presenceId}/close:
    post:
      operationId: closePresence
      summary: Close employee presence
      parameters:
        - in: path
          name: employeeId
          required: true
          schema:
            type: integer
            format: int64
        - in: path
          name: presenceId
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClosePresenceRequest'
      responses:
        '200':
          description: Presence closed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PresenceResponse'
        '400':
          description: Invalid request
        '404':
          description: Presence not found
        '409':
          description: Presence already closed
...
components:
  schemas:
...
    EmployeeResponse:
      type: object
      required:
        - id
        - ruleSystemCode
        - employeeNumber
        - firstName
        - lastName1
        - status
      properties:
        id:
          type: integer
          format: int64
        ruleSystemCode:
          type: string
        employeeNumber:
          type: string
        firstName:
          type: string
        lastName1:
          type: string
        lastName2:
          type: string
          nullable: true
        preferredName:
          type: string
          nullable: true
        status:
          type: string
    CreatePresenceRequest:
      type: object
      required:
        - companyCode
        - entryReasonCode
        - startDate
      properties:
        companyCode:
          type: string
          maxLength: 4
        entryReasonCode:
          type: string
          maxLength: 30
        exitReasonCode:
          type: string
          maxLength: 30
          nullable: true
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
          nullable: true
    ClosePresenceRequest:
      type: object
      required:
        - endDate
      properties:
        endDate:
          type: string
          format: date
        exitReasonCode:
          type: string
          maxLength: 30
          nullable: true
    PresenceResponse:
      type: object
      required:
        - id
        - employeeId
        - presenceNumber
        - companyCode
        - entryReasonCode
        - startDate
      properties:
        id:
          type: integer
          format: int64
        employeeId:
          type: integer
          format: int64
        presenceNumber:
          type: integer
          format: int32
        companyCode:
          type: string
        entryReasonCode:
          type: string
        exitReasonCode:
          type: string
          nullable: true
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
          nullable: true
...